+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Version 5.3
 - PVOnline connectivity improvements
 - Fix a bug preventing set formulas to internal use customer
+- Add stock quantities in the ingredients page
 
 ### Version 5.2
 - Extend API to include allergen status