+++ releasenotes.md
@@ -2,3 +2,4 @@ Whats New:
 ----------------------
 - PVOnline connectivity improvements
 - Fix a bug preventing set formulas to internal use customer
+- Add stock quantities in the ingredients page